refactor: move get_preverb_from_head_breakdown -> search.py
